% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaned_tidy.R
\name{cleaned_tidy}
\alias{cleaned_tidy}
\title{Additional formatting options for a broom::tidy() data.frame}
\usage{
cleaned_tidy(x, hun = FALSE)
}
\arguments{
\item{x}{model or tidied data.frame}

\item{hun}{Hungarian output? (default FALSE)}
}
\description{
Additional formatting options for a broom::tidy() data.frame
}
\examples{
m <- broom::tidy(lm(Sepal.Length ~ Petal.Width, iris))
cleaned_tidy(m)

m <- lm(Sepal.Length ~ Petal.Width, iris)
cleaned_tidy(m, hun = TRUE)
}
