% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorised_datatable.R
\name{colorised_datatable}
\alias{colorised_datatable}
\title{colorised_datatable}
\usage{
colorised_datatable(.data, .cols = NULL, type = "bg-color", ...)
}
\arguments{
\item{.data}{A data.frame.}

\item{.cols}{Columns to colorise (by default all the numrical columns).}

\item{type}{"bar" or "bg-color" (default).}
}
\description{
Colorise  or draw column for all numeric or selected columns (chr vector) of a data.frame
}
\examples{
colorised_datatable(iris, "Sepal.Length", type = "bar")
colorised_datatable(iris)
}
