% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer.R
\name{cramer}
\alias{cramer}
\title{cramer}
\usage{
cramer(x, keep.numeric = F, dlimit = NULL)
}
\arguments{
\item{x}{Data.frame}

\item{keep.numeric}{Logical; Should use numeric values as factors and keep them? (FALSE)}

\item{dlimit}{Integer; If a numeric column has more unique values than this limit, then it will be eliminated(NULL)}
}
\value{
Data.frame, which has 3 column: the compared 2 variables & the value of cramer indicator. \code{overview_print}
}
\description{
Calculates all the pairwised cramer values in a data.frame.
}
\examples{
data(Affairs, package = "AER")
cramer(Affairs)
cramer(Affairs, keep.numeric = T, dlimit = 7)
}
