% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pb.R
\name{create_pb}
\alias{create_pb}
\title{Create progress bar.}
\usage{
create_pb(x, .message = TRUE)
}
\description{
Create progress bar.
}
\examples{
granatlib::create_pb(100)
for (i in 1:100) {
  Sys.sleep(.1)
  pb$tick()
}
}
