% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pin.R
\name{create_pin}
\alias{create_pin}
\title{Create a secret pin}
\usage{
create_pin(onedrive_folder = NULL)
}
\arguments{
\item{onedrive_folder}{Name of the folder to use}
}
\description{
Create a secret pin
}
