% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dall_e.R
\name{dall_e}
\alias{dall_e}
\title{Download a Dall-e image.}
\usage{
dall_e(promt, folder.path = "figures/", size = "1024x1024")
}
\arguments{
\item{promt}{(string) The prompt for the image.}

\item{folder.path}{(string) The folder where to save the image. Finish it with "/".}

\item{size}{(string) The size of the generated images. Must be one of 256x256, 512x512, or 1024x1024 (default).}
}
\description{
This function uses {openai} to generate an image with Dall-e, but downloads it into a specified folder. The snippet for markdown usage also added to the clipboard.
}
\examples{
dall_e("R programmer writing a new package.")
}
