% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depedency_check.R
\name{depedency_check}
\alias{depedency_check}
\title{depedency_check}
\usage{
depedency_check(file_name)
}
\description{
Check the dependencies of a function and compare it with the import field. Also possible to rewrite the code with explicit function calls.
}
