% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify.R
\name{dummify}
\alias{dummify}
\title{dummify}
\usage{
dummify(.data, .col = NULL, sep = "_")
}
\description{
Dummify all non-numeric or dplyr::selected columns (chr vector) of a data.frame
}
\examples{
dummify(ggplot2::diamonds, "cut")

dummify(ggplot2::diamonds)
}
