% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_diff.R
\name{environment_diff}
\alias{environment_diff}
\title{environment_diff}
\usage{
environment_diff()
}
\description{
Check the difference of the environment between the two state of function call
}
\examples{
d <- 2
e <- 3
environment_diff()
rm(d)
e <- 5
a <- 3
environment_diff()
}
