% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_country_map.R
\name{eu_country_map}
\alias{eu_country_map}
\title{eu_country_map}
\usage{
eu_country_map(x, text_size = 1.5, label_fun = function(x) x)
}
\description{
Country map.
}
\examples{
eurostat::get_eurostat("ei_lmhu_m") |>
  group_by(geo) |>
  slice_max(time) |>
  filter(indic == "LM-UN-M-TOT", s_adj == "SA") |>
  select(geo, values) |>
  eu_country_map()

}
