% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_publication.R
\name{find_publication}
\alias{find_publication}
\title{Find publications with dplyr::dplyr::filters based on PorP exported file}
\usage{
find_publication(file, n = NULL, open = FALSE, ...)
}
\arguments{
\item{file}{csv file}

\item{n}{Number of articles}

\item{open}{Open in your default browser? if FALSE, then data.frame is returned}
}
\description{
Find publications with dplyr::dplyr::filters based on [PorP](https://harzing.com/resources/publish-or-perish) exported file (also sci-mago dplyr::dplyr::filters)
}
\examples{
find_publication("https://gist.githubusercontent.com/MarcellGranat/4ed653c8e655d4ebabaa3071fc7b50a0/raw/5aef770bce91b661744ad8c8b1aed56643f795cf/daily-inflation-online.csv", n = 10, FALSE, year > 2015)
}
