% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tag.R
\name{geom_tag}
\alias{geom_tag}
\title{geom_tag}
\usage{
geom_tag(
  x,
  y,
  label = "label",
  color = "cyan4",
  shape = 16,
  nudge_x = -0.5,
  nudge_y = -0.5,
  point_size = 3,
  label_size = 3
)
}
\description{
highlight a datapoint on the figure with label.
}
