% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtf_idf.R
\name{ggtf_idf}
\alias{ggtf_idf}
\title{ggtf_idf}
\usage{
ggtf_idf(
  x,
  stem_function = function(x) corpus::stem_snowball(x, "hu"),
  n = 10,
  language = "hu",
  drop_words = stopwords::stopwords("hu"),
  min_bigram_n = 1,
  drop_numbers = TRUE
)
}
\arguments{
\item{x}{A data.frame containing one fct column and one column of text. Levels of the fct column gives the categories for the TF-IDF calculation, but only the first and the last level will appear on the plot}
}
\value{
Ggobject \code{overview_print}
}
\description{
Draws a graph of the bigrams, which contain at least one of the words with the highest TF-IDF value.
}
\examples{
plot3D()
}
