% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl.R
\name{hl}
\alias{hl}
\title{hl}
\usage{
hl(output_html, output_latex)
}
\value{
one of the given outputs based on the environment
}
\description{
Manage the different behaviour of html & kable output. Short name (Html-Latex) for markdown inline formatting:  r hl(x, y).
}
\examples{
library(ggplot2)
p <- ggplot(iris, aes(Sepal.Length, Petal.Length)) + geom_point()
hl(plotly::ggplotly(p), p)
}
