% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mice.R
\name{impute_mice}
\alias{impute_mice}
\title{impute_mice}
\usage{
impute_mice(.data, method = "rf", ...)
}
\arguments{
\item{.data}{Data.frame with missing values}

\item{method}{method to impute (rf by default)}

\item{...}{Additional arguments for mice}
}
\value{
Imputed data.frame \code{overview_print}
}
\description{
using functions from package mice to impute missing data and return the data.frame
}
\examples{
create_pb()
}
