% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Display message.}
\usage{
info(msg = "", type = "info", add_time = TRUE)
}
\arguments{
\item{msg}{The message to display. Use {} for glue type interpolating.}

\item{type}{Character. One of "info" (default)/"ok"/"warning". This will determine the color and the symbol.}

\item{add_time}{Logical. Shoudl the time printed? (default TRUE)}
}
\description{
Print a nicely formatted message, optionally with system time.
}
\examples{
x = 3
info("The value of x is {x}.", "info", add_time = TRUE)
}
