% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_output.R
\name{kable_output}
\alias{kable_output}
\title{kable_output}
\usage{
kable_output(
  .data,
  align = NULL,
  round_digits = NULL,
  plus_sign = FALSE,
  hun = FALSE,
  same_digits = FALSE,
  verticals = FALSE,
  .keep_lines = c(1, 2, -1),
  force_large = FALSE,
  caption = NULL,
  text_contained = NULL,
  ...
)
}
\value{
knitr::kable output \code{overview_print}
}
\description{
IF not render then show the table and copy it to clipboard in latex format. Round & decimal mark also managed
}
\examples{
kable_output(iris)
}
