% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_source.R
\name{logged_source}
\alias{logged_source}
\title{Source, but created a .log file in _log folder.}
\usage{
logged_source(script)
}
\arguments{
\item{script}{Script to run.}
}
\description{
Rewrites # --- titles in the script with cli messages.
}
