% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_insert.R
\name{md_insert}
\alias{md_insert}
\title{md_insert}
\usage{
md_insert(
  x,
  text_contained = NULL,
  asis = TRUE,
  fig_captions = NULL,
  tab_captions = NULL,
  todo_contained = NULL
)
}
\description{
Insert a markdown (plain text) code into the rmd file. Use it with inline code!
}
\examples{
`r md_insert("how_to.md")`
}
