% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_times_try.R
\name{n_times_try}
\alias{n_times_try}
\title{Call TryCatch n times with given sleep times.}
\usage{
n_times_try(
  .exp,
  sleep_times = c(3, 3, 3),
  otherwise = NULL,
  print_message = FALSE
)
}
\arguments{
\item{.exp}{Expression}

\item{otherwise}{Value to return if all the tries resulted an error.}

\item{print_message}{Should print the message (default FALSE).}

\item{sleep_time}{Numeric vector given the seconds to wait for next try.}
}
\description{
Mainly suggested for web scraping when reaching a given domain in a short time leads to errors.
}
\examples{
v <- list("a", "B", 2) # only 3rd element can be added to another numeric
i = 0
n_times_try({
i = i + 1
v[[i]] + 3
})

}
