% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{plot3D}
\usage{
plot3D(
  f = function(x, y) x^0.2 * y^0.2,
  x.min = 0,
  y.min = 0,
  x.max = 10,
  y.max = 10,
  n = 100,
  low = "blue",
  high = "red",
  title = NULL,
  legend.title = "f",
  xlab = "x",
  ylab = "y"
)
}
\arguments{
\item{f}{The function, which will be plotted in 3D (function(x, y) x^0.2*y^0.2)}

\item{x.min}{Lower limit of x-axis (0)}

\item{y.min}{Lower limit of y-axis (0)}

\item{x.max}{Upper limit of x-axis (10)}

\item{y.max}{Upper limit of y-axis (10)}

\item{n}{Number of points in rows and columns (100)}

\item{low}{Colour of the lowest value ("blue")}

\item{high}{Colour of the highest value ("red")}

\item{title}{Title of the plot (f as expression)}

\item{legend.title}{Title of the legend ("f")}

\item{xlab}{Title of the x-axis ("x")}

\item{ylab}{Title of th y-axis ("y")}
}
\value{
3D figure based on rayshader. \code{overview_print}
}
\description{
Plot a function with 2 input values in 3D based on ggplot2 & rayshader package.
}
\examples{
plot3D()
}
