% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_str.R
\name{repo_str}
\alias{repo_str}
\title{Structure the repository based on Marcell's preferences.}
\usage{
repo_str(board = TRUE, chatgpt = TRUE)
}
\arguments{
\item{board}{logical. Should a script for {pins} added?}

\item{chatgpt}{logical. Should chatGPT add comments to the scripts?}
}
\description{
Adds README, utils, board and .Rprofile files and updates them.
}
\examples{
repo_str()
}
