% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowid.R
\name{rowid}
\alias{rowid}
\title{rowid}
\usage{
rowid(prefix = "id", suffix = "")
}
\arguments{
\item{prefix}{Character. Prefix of the generated id. Default is "id".}

\item{suffix}{Character. Suffix of the generated id. Default is "".}
}
\description{
Extension of dplyr::row_number(). This generates 001 if there are more than 100 observations.
}
\examples{
tibble::tibble(x = 1:1000) |>
 dplyr::mutate(id = rowid())
}
