% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_based.R
\name{sort_based}
\alias{sort_based}
\title{Sort the elements based on a results of a function called on them.}
\usage{
sort_based(x, .f)
}
\arguments{
\item{x}{A vector or list.}

\item{.f}{Called function.}
}
\description{
Sort the elements based on a results of a function called on them.
}
\examples{
sort_based(c("a22", "b33", "c15"), parse_number)

}
