% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitted_mutate.R
\name{splitted_mutate}
\alias{splitted_mutate}
\title{splitted_mutate}
\usage{
splitted_mutate(.data, ..., split_number = 100, .keep = NULL, .progress = TRUE)
}
\arguments{
\item{.data}{Data frame.}

\item{...}{arguments to pass mutate(...).}

\item{split_number}{Number of parts to split the data.frame before evaluation.}

\item{.keep}{Variables to keep after eval the expression. Original columns are keept automaticly.}

\item{cores}{Number of cores to use. If cores = 1 (default) then evaluation is made by simple for loop.}
}
\value{
.data \code{overview_print}
}
\description{
Splits the data.frame into separeta parts and evaluates it one-by-one. Useful for saving memmory during large computations.
}
\examples{

iris_samples <- tibble(data = replicate(100, sample_n(iris, 50), simplify = FALSE))

iris_samples

iris_samples \%>\%
 splitted_mutate(
   fit = map(data, lm, formula = Sepal.Length ~ Sepal.Width),
   coefs = map(fit, "coefficients"),
   coef1 = map_dbl(coefs, 1),
   coef2 = map_dbl(coefs, 2),
   .keep = c("coef1", "coef2") # single core
 )

iris_samples \%>\%
 splitted_mutate(
   fit = map(data, lm, formula = Sepal.Length ~ Sepal.Width),
   coefs = map(fit, "coefficients"),
   coef1 = map_dbl(coefs, 1),
   coef2 = map_dbl(coefs, 2),
   .keep = c("coef1", "coef2"),
   cores = 6 # multicore evaluation
 )

}
