% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gR.R
\name{theme_gR}
\alias{theme_gR}
\title{theme_gR}
\usage{
theme_gR(base_theme = NULL)
}
\arguments{
\item{base_theme}{A theme to use}
}
\description{
A theme that is conditionally dark or minimal (by default). Dark if you use it in RStudio.
}
\examples{
ggplot(iris, aes(Sepal.Length, Petal.Width)) +
  geom_point() +
  theme_gR()

}
