% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot_summarise.R
\name{tot_summarise}
\alias{tot_summarise}
\title{tot_summarise}
\usage{
tot_summarise(x, ..., total_name = "Total")
}
\arguments{
\item{x}{A grouped data.frame}

\item{...}{Arguments pass to summarise(...)}

\item{total_name}{Label for the total row. By default it is "Total"}
}
\value{
A grouped data.frame \code{overview_print}
}
\description{
Add a total as a new level to the second grouping column. (e.g.: total rate in each industry and in all (2nd col) by years (1st col))
}
\examples{
iris |>
 dplyr::group_by(Species) |>
 tot_summarise(mean(Sepal.Length))

}
