% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification.R
\name{notification}
\alias{notification}
\title{Send a notification on Mac.}
\usage{
notification(..., sound = FALSE)
}
\arguments{
\item{...}{Message.}

\item{sound}{Should play a sound (default FALSE).}
}
\description{
Sends a notification on Mac. Retrives a sound only on Windows.
}
