% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_str.R
\name{repo_str}
\alias{repo_str}
\title{Structure the repository based on Marcell's preferences.}
\usage{
repo_str(board = TRUE)
}
\description{
Adds README, utils, board and .Rprofile files and updates them.
}
\examples{
repo_str()
}
